Palette now uses universal palettes
The shader now uses a scriptable object with the lists of colours on them. This is used on all objects that use the shader. When you make an instance it lets you use the 4th colour which you can edit in the editor window opened with the "open editor" in the shader inspector.